refactor(#255): extract test_support.rs (make_output, make_output_full, load_fixture)
Move the inline #[cfg(test)] pub(crate) mod test_support block from mod.rs
into a new crates/rskim/src/cmd/test_support.rs file. mod.rs retains:

  #[cfg(test)]
  pub(crate) mod test_support;

The file reference keeps the same #[cfg(test)] gating — test_support.rs is
compiled only during test builds. All callers (super::test_support::make_output,
etc.) continue to work unchanged via the module re-declaration.

mod.rs drops from 494 to 421 lines. All 3,640+ tests pass.
